% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_coefficient.R
\name{binomial_coefficient}
\alias{binomial_coefficient}
\title{binomial coefficent}
\usage{
binomial_coefficient(n, k)
}
\arguments{
\item{n}{integer; must be larger or equal to k.}

\item{k}{integer; must be larger or equal to 0.}
}
\description{
a function that calculates the binomial coefficent
}
\details{
the binomial coefficent is defined as \deqn{n!/(k! * (n-k)!)} where
! denotes the factorial operator. n and k must be positive integers satisfying
\eqn{n >= k}.
}
\seealso{
\url{https://en.wikipedia.org/wiki/Binomial_coefficient}
}
