% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_palindrom.R
\name{is_palindrom}
\alias{is_palindrom}
\title{checks if a string is a palindrom}
\usage{
is_palindrom(str)
}
\description{
checks if a string is a palindrom
}
\seealso{
\url{https://en.wikipedia.org/wiki/Palindrome}
}
