% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiples_of_3_and_5.R
\name{multiples}
\alias{multiples}
\title{Gives the sum of all natural numbers below \code{x} that are multiples of 3 or 5.}
\usage{
multiples(x)
}
\arguments{
\item{x}{A natural number.}
}
\value{
The sum of all natural numbers below \code{x} that are multiples of 3 or 5.
}
\description{
Gives the sum of all natural numbers below \code{x} that are multiples of 3 or 5.
}
