% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permute_vector}
\alias{permute_vector}
\title{calculate all permutations of a vector x}
\usage{
permute_vector(x)
}
\arguments{
\item{x}{a vector}
}
\value{
matrix; each column is one unqiue permutation of x
}
\description{
calculate all permutations of a vector x
}
\details{
this is a naiive recursive implementation
}
